requirements.txt
matplotlib
setuptools
requests
pandas
PyYAML